+++ src/SFA.DAS.Campaigns.UITests/Helpers/CampaignsStepsHelper.cs
@@ -1,5 +1,7 @@
 ﻿using SFA.DAS.Campaigns.UITests.Project.Tests.Pages;
 using SFA.DAS.Campaigns.UITests.Project.Tests.Pages.Apprentices;
+using SFA.DAS.Campaigns.UITests.Project.Tests.Pages.Employer;
+using SFA.DAS.Campaigns.UITests.Project.Tests.Pages.Influencers;
 
 namespace SFA.DAS.Campaigns.UITests.Helpers
 {
