+++ src/SFA.DAS.Campaigns.UITests/Project/Tests/Pages/CampaingnsHeaderBasePage.cs
@@ -1,4 +1,6 @@
 ﻿using SFA.DAS.Campaigns.UITests.Project.Tests.Pages.Apprentices;
+using SFA.DAS.Campaigns.UITests.Project.Tests.Pages.Employer;
+using SFA.DAS.Campaigns.UITests.Project.Tests.Pages.Influencers;
 
 namespace SFA.DAS.Campaigns.UITests.Project.Tests.Pages;
