+++ src/SFA.DAS.Campaigns.UITests/Project/Tests/Pages/Employer/EmployerBasePage.cs
@@ -1,0 +1,5 @@
+﻿namespace SFA.DAS.Campaigns.UITests.Project.Tests.Pages.Employer;
+
+public abstract class EmployerBasePage(ScenarioContext context) : HubBasePage(context)
+{
+}
