+++ src/SFA.DAS.Campaigns.UITests/Project/Tests/StepDefinitions/TraceStartAndStopSteps.cs
@@ -1,36 +1,35 @@
-﻿namespace SFA.DAS.Campaigns.UITests.Project.Tests.StepDefinitions
+﻿namespace SFA.DAS.Campaigns.UITests.Project.Tests.StepDefinitions;
+
+[Binding]
+public class TraceStartAndStopSteps(ScenarioContext context)
 {
-    [Binding]
-    public class TraceStartAndStopSteps(ScenarioContext context)
+    [Given("the trace is started")]
+    public async Task GivenTheTraceIsStarted()
     {
-        [Given("the trace is started")]
-        public async Task GivenTheTraceIsStarted()
+        var driver = context.Get<Driver>();
+
+        await driver.BrowserContext.Tracing.StartAsync(new()
         {
-            var driver = context.Get<Driver>();
+            Title = context.ScenarioInfo.Title,
+            Screenshots = true,
+            Snapshots = true
+        });
+    }
 
-            await driver.BrowserContext.Tracing.StartAsync(new()
-            {
-                Title = context.ScenarioInfo.Title,
-                Screenshots = true,
-                Snapshots = true
-            });
-        }
+    [Given("the trace is stopped")]
+    public async Task GivenTheTraceIsStopped()
+    {
+        var driver = context.Get<Driver>();
 
-        [Given("the trace is stopped")]
-        public async Task GivenTheTraceIsStopped()
+        var tracefileName = $"TRACEDATA_{DateTime.Now:HH-mm-ss-fffff}.zip";
+
+        var tracefilePath = $"{context.Get<ObjectContext>().GetDirectory()}/{tracefileName}";
+
+        await driver.BrowserContext.Tracing.StopAsync(new()
         {
-            var driver = context.Get<Driver>();
+            Path = tracefilePath
+        });
 
-            var tracefileName = $"TRACEDATA_{DateTime.Now:HH-mm-ss-fffff}.zip";
-
-            var tracefilePath = $"{context.Get<ObjectContext>().GetDirectory()}/{tracefileName}";
-
-            await driver.BrowserContext.Tracing.StopAsync(new()
-            {
-                Path = tracefilePath
-            });
-
-            TestContext.AddTestAttachment(tracefilePath, tracefileName);
-        }
+        TestContext.AddTestAttachment(tracefilePath, tracefileName);
     }
 }
